Defer evaluation of optional properties.
The cloneGhPages task requires a GitHub username and password, which should be provided as Gradle properties. If those properties are not provided, the task should fail. However, if the user doesn't execute that task then the properties aren't required. Unfortunately, the configuration of all tasks happens up-front, at which point an exception is thrown if the GitHub username and password are not set, regardless of which tasks the user is running.

This fix uses the lazily-evaluated GString trick to defer property evaluation until such time as the task is executed.